fix: duplicate log message parsing
