+++ lua/neogit/lib/git/rebase.lua
@@ -1,83 +1,17 @@
 local util = require("neogit.lib.util")
 local logger = require("neogit.logger")
 local client = require("neogit.client")
+local log = require("neogit.lib.git.log")
 
 local M = {}
 
 local commit_header_pat = "([| *]*)%*([| *]*)commit (%w+)"
 
-local function is_new_commit(line)
-  local s1, s2, oid = line:match(commit_header_pat)
-
-  return s1 ~= nil and s2 ~= nil and oid ~= nil
-end
-
--- NOTE: this is duplicated
-local function parse(raw)
-  local commits = {}
-  local idx = 1
-
-  local function advance()
-    idx = idx + 1
-    return raw[idx]
-  end
-
-  local line = raw[idx]
-  while line do
-    local commit = {}
-    local s1, s2
-
-    s1, s2, commit.oid = line:match(commit_header_pat)
-    commit.level = util.str_count(s1, "|") + util.str_count(s2, "|")
-
-    local start_idx = #s1 + #s2 + 1
-
-    local function ladvance()
-      local line = advance() or ""
-      return line and line:sub(start_idx + 1, -1) or nil
-    end
-
-    do
-      local line = ladvance() or ""
-
-      if vim.startswith(line, "Merge:") then
-        commit.merge = line:match("Merge:%s*(%w+) (%w+)")
-
-        line = ladvance()
-      end
-
-      commit.author_name, commit.author_email = line:match("Author:%s*(.+) <(.+)>")
-    end
-
-    commit.author_date = ladvance():match("AuthorDate:%s*(.+)")
-    commit.committer_name, commit.committer_email = ladvance():match("Commit:%s*(.+) <(.+)>")
-    commit.committer_date = ladvance():match("CommitDate:%s*(.+)")
-
-    advance()
-
-    commit.description = {}
-    line = advance()
-
-    while line and not is_new_commit(line) do
-      table.insert(commit.description, line:sub(start_idx + 5, -1))
-      line = advance()
-    end
-
-    if line ~= nil then
-      commit.description[#commit.description] = nil
-    end
-
-    table.insert(commits, commit)
-  end
-
-  return commits
-end
-
 function M.commits()
   local git = require("neogit.lib.git")
   local output = git.cli.log.format("fuller").args("--graph").call_sync(true).stdout
 
-  return parse(output)
+  return log.parse(output)
 end
 
 local a = require("plenary.async")
@@ -120,6 +54,7 @@ function M.update_rebase_status(state)
   local _, stat = a.uv.fs_stat(root .. "/.git/rebase-merge")
   local rebase_file = nil
 
+  -- Find the rebase progress files
   if stat then
     rebase_file = root .. "/.git/rebase-merge"
   else
