+++ lua/neogit/lib/util.lua
@@ -8,6 +8,20 @@ local function map(tbl, f)
   return t
 end
 
+---@param tbl any[]
+---@param f fun(v: any) -> any|nil
+---@return any[]
+local function filter_map(tbl, f)
+  local t = {}
+  for _, v in ipairs(tbl) do
+    v = f(v)
+    if v ~= nil then
+      table.insert(t, v)
+    end
+  end
+  return t
+end
+
 ---@param value number
 ---@param min number
 ---@param max number
@@ -164,6 +178,7 @@ return {
   clamp = clamp,
   slice = slice,
   map = map,
+  filter_map = filter_map,
   range = range,
   filter = filter,
   str_right_pad = str_right_pad,
